Cleanup FEN parser
